requirements.txt
asgiref==3.8.1
click==8.2.0
coverage==7.8.0
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-import-export==4.3.7
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environ==1.0
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
rq==2.3.3
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
uritemplate==4.1.1
